refactor: move source detection helpers to separate module
Extract isRedHatSource and isRhlwSource from analysis.ts into
sourceDetection.ts to break the circular dependency between
extension.ts and analysis.ts.

Implements TC-5425
